Support different serialization techniques for the cache.
Sometimes, content stored in the cache has some meaning to it (it might
represent a data structure). Therefore, it is useful to have a way to
tell the caching engine how to persist the cached content to a storage
as well as how to build it back when reading from the storage.

This adds the concept of "cache serializers". These are classes that
implement a `encode` and a `decode` method and are able to implement
custom parsing to content stored in the cache.

The default serializer is the `text` serializer, that just stores
content as strings. However, there is also support JSON serialization
for the caching of more complicated, key-value data structures.

Supplier clients that need to deal with customized formats can implement
their own serializers that wrap the cached content in whatever object is
convenient.