Make sure the return type of `extract_rates` is consistent.
Previously, the `extract_rates` in JTB's response parser could return a
Result object wrapping either an array of rate plans, or a +RatePlan+
instance indicating the unit is not available on the selected dates.

This inconsistency lead to a bug where a collection of rates was
expected when actually a +RatePlan+ was returned.

This fixes the issue by returning an empty list of rate plans when the
unit is not available, and letting the caller pass the error upstream on
that case, considering the unit not to be available, making the return
type of `extract_rates` consistent.