Fix time-of-check, time-of-use issue when updating the cache.
Depending on the order of the events when writing to the cache, it could
be possible that, when checking a given key was cached or not, another
process wrote to the cache using the same key. This caused a unique
constraint violation error to the original call.

This commit fixes the issue by recovering from such scenarios and
attempting an update in such cases.