Support for app-specific list of required env variables.
With the addition of a `web` app, it is desired to have a list of
per-app required environment variables, since variables required for one
app might not be required for the other. However, there must still be
support for a list of globally required environment variables, no matter
the app. This introduces these changes.

It also adds the concept of `Concierge.app`, which tells which app is
currently running by inspecting the content of the `CONCIERGE_APP`
variable, which must be defined on boot time.