Do not cause master process to die if child was manually killed.
If a child is manually killed before the master was sent a deadly
signal, the master process would crash since it would try to send a
signal to a non-existent process. This fixes adds a fix to such scenario
by recovering from such an error and attempting to terminate the other
child processes.