Fix content type identification by normalising header names.
Previously, JSON content types were not being recognised since the
Network Response partial was checking for a `Content-Type` header in the
response. However, some responses send the header information all in
lower case characters, and the identification of the content type should
still work in such scenarios.

This normalizes the headers coming from the request context before
comparisons are made so that the identification is more accurate.