Ignore price_check events.
This Roomorama webhook, even though currently enforced for a property to
be considered IB, follows a pattern that does not fit the current
integrations model: it is sent asynchronously and expects the supplier
to check if the price given is correct, and update availabilities in
case it is not.

Since Concierge suppliers need to subscribe to this event for
instant-confirmation, the previous result was causing two `quote` calls
to be performed on a single price check.

This ignores any incoming webhook carrying the `price_check` event,
knowing that the price will properly be calculated via the
`quote_instant` webhook.