Allow context tracking to be disabled/enabled without data loss.
For synchronisation workers, it is sometimes not desired to have context
tracking, since it potentially involves HTTP requests to fetch large
files, making the persistence and analysis of such data pointless.

However, when performing calls to Roomorama to publish/update
properties, it is required to have context tracking enabled so that any
errors that happen in the process can be persisted and analysed later.

This introduces `enable!` and `disable!` methods on `Concierge::Context`
instances, allowing the caller to temporarily disable context tracking,
and allowing it to be enabled back again without loss of previously
tracked events.