Allow Concierge::SafeAccessHash to be serialized to JSON.
Prior to this commit, loading an `ExternalError` from
`ExternalErrorRepository`, making changes to it and then calling
`.update` would draw an error since `dump` expects a Ruby `Hash` and the
`load` method parses the JSON from the database into a
`Concierge::SafeAccessHash` instance.

The mismatch causes the failure in serializing the `context` column to
JSON when updating records. This calls `.to_h` on the received instance
so that  `dump` and `load` are consistent.