Do not send calendar fields if they are the same as Roomorama defaults.
Roomorama's calendar API assumes calendars dates to allow checking-in
and checking-out as well as assuming reasonable defaults for
weekly/monthly prices if the supplier does not provide them.

This removes such fields from the payload sent to Roomorama if the
supplier does not provide them, significantly reducing the payload sent
in such cases.