Include artificial /checkout endpoint for all suppliers.
Currently, Roomorama enforces the presence of a `checkout_instant`
event, triggered when the user is interested in booking a property, but
has not paid yet.

With the flow we have been working with on Concierge, the existence of
this event is useless and we commonly have pointing the
`checkout_instant` to the `quote` endpoint just so that things "work".
The downside of this is that the `quote` call performs API calls to
suppliers, and can contribute to our rate limit.

This includes an aritificial `/checkout` endpoint, which is always
successful. Such endpoint can be configured for every Concierge
integration, until we migrate the webhooks flow on Roomorama to not
depend on this event, which is a more involved change.