Do not mark worker as `running` when scheduling.
This changes the `Workers::Scheduler` class to *not* change the
background worker status to `running` when it is scheduled. That is
because the queue processor already manages the worker status, and
therefore there was a duplication. Also, enqueueing a job for processing
does not mean the worker is actually running yet.

Fixes an issue introduced by b04b1805550fa46458ba8fa69264b2a846a2f1e4