Ability to include unit availabilities on Roomorama::Calendar.
The calendar of availabilities can also be updated on Roomorama on a
per-unit basis, for multi-unit property provider.

This includes a `add_unit` method to `Roomorama::Calendar` which expects
another instance of `Roomorama::Calendar` to be passed, representing the
availabilities calendar of a unit of the parent property.

In that format, the `to_h` method will then serialise the availabilities
calendar for the property and associated units.