Do not attempt to update calendar for properties not previously synced.
It is common for the calendar endpoint of supplier APIs to be different
from that to get property data. Therefore, it is possible for APIs to
return the availabilities calendar of properties that were not
previously synchronised.

Since this is a possible scenarion across different suppliers, this
changes the `Workers::CalendarSynchronisation` class to skip updating
the calendar (that is, making an API call to Roomorama) in case the
property identifier is unknown.

In case the property is indeed created by Concierge on a later run of
the `metadata` worker, then the next `calendar` worker run should update
its calendar as usual.