Rename Ciirus availabilities worker to reflect event name.
Even though the API call on Roomorama is called `update_calendar` we
ended up deciding to use the name `availabilities` for the worker type
that is supposed to synchronise availabilities between suppliers and
Roomorama. The intention was to standardise naming, since there was a
point where both names (`availabilities` and `calendar` were being used
to represent the same thing.)

However, we forgot to change the name of the event being listened to by
the Ciirus worker, which was still expecting a `calendar.Ciirus` event
to be sent. This changes the file name to reflect the naming update, as
well as fixes the event name on the `Concierge::Announcer` call at the
end of the Ciirus availabilities worker implementation.