Do not attempt to call `to_h` in case the stored result is an Array.
`Array#to_h` is equivalent to `Hash[<array>]`, and therefore is a very
particular implementation. Kigo's flow introduced arrays on the cache
for the first time, and calling `to_h` on the resulting array lead to
problems.

This changes the view to avoid calling `to_h` in case the result is an
array.