Increase queue timeout to 12h.
With the addition of Poplidays and AtLeisure, which need to process
large amounts of properties (20k-30k) in a single host account, we need
to increase the timeout allowed for such workers to complete.

We are probably hitting the timeout limit very soon. If this starts to
become a bottleneck, we'd better think of a way to complete the
synchronisation within a reasonable time (i.e., consistently < 12h.)