Proceed successfully when there are no identifiers to be updated.
When there are no identifiers to be updated for a calendar sync, the
Kigo calendar worker would fail since it assumed at least one property
should be updated. This fixes this behaviour by saving a synchronisation
process with 0 records processed in such occurrences.

It also renames `HostCheck#check` to `HostCheck#active?` for a better
mapping of whether the host is active (in which case the method returns
`true`) or not.