Remove 12h timeout requirement.
Suppliers such as Poplidays and AtLeisure provide a large number of
properties all listed under the same host. For that reason, the existing
12h timeout enforced on workers execution cannot be met. They have been
frequently timing out.

The timeout is there in order to reflect the current VisibilityTimeout
setting on the underlying SQS queue. When the worker times out, the
message is not deleted from the queue, but is kept there for later
processing.

This changes the logic of queue polling by instantly removing the
message from the queue as soon as a worker picks it up. As the
synchronisation tasks are handled by
`Workers::Processor::BackgroundWorker`, the `background_worker`
counterpart row on the database will have its `next_run_at` column
updated after the worker is executed. Therefore, on failure or success,
the chain is guaranteed to continue. The enforcement, therefore, is
unnecessary.

After this is deployed, the VisibilityTimeout setting can be changed to
a low value (e.g., 10 minutes), so as to guarantee that the worker that
picks up the message has enough time to delete it from the queue. And
therefore no two workers will be processing the same message.