Mark workers as queued to avoid reprocessing.
This addresses a specific series of events that have been happening
during background worker processing:

1. BackgroundWorker is idle, and therefore is put in the queue by the
scheduler.
2. The queue is full, and the workers are all busy processing messages
in the queue.
3. Ten minutes later, the scheduler kicks in again. Noticing that the
background worker queued on step number 1 is still not processed and
with a `next_run_at` column in the past, it enqueues it again. As a
result, there are now two messages in the queue to synchronise the same
host, when there should be only one.

This deals with the situation described above by marking the workers as
`queued` whenever they are put into SQS. Therefore, it will then move to
`running` and `idle` again to complete the cycle. Most importantly, the
scheduler will no longer add it to the queue if the previous run already
added the same worker.

With this change, the number of messages in the queue is expected to
decrease, as we remove the duplications.